Canonical CI: grouped-tests.yml + root test/test_groups.toml
Convert the root test workflow to the canonical thin caller of
SciML/.github grouped-tests.yml@v1, with the version/group/OS matrix
declared once in test/test_groups.toml.

Category B refactor (Aqua/JET previously ran inline in runtests.jl):
- Functional interface tests moved to test/core_tests.jl (GROUP Core/All),
  alongside the AllocCheck allocation tests.
- Aqua + JET consolidated into test/qa.jl, gated on GROUP=="QA", running
  in an isolated test/qa/Project.toml environment (Aqua, JET, Test,
  LinearAlgebra, Statistics + SurrogatesBase via [sources] path="../..").
- runtests.jl is now GROUP-dispatched: QA activates/instantiates the qa
  env and develops the root package before including qa.jl.
- Aqua/JET removed from the root [extras]/[targets].test/[compat] now that
  they live only in the QA env (keeps QA tooling out of reverse-dep
  resolution).

test/test_groups.toml:
- [Core] on [lts, 1, pre] across [ubuntu-latest, macos-latest,
  windows-latest] (preserves the old OS matrix).
- [QA] on [lts, 1] (ubuntu).

Tests.yml is now a thin caller; on:/concurrency: preserved verbatim. The
standalone alloccheck job is removed (its coverage is subsumed by Core).
All other workflows untouched.
