+++ test/qa.jl
@@ -1,4 +1,10 @@
-using SurrogatesBase, Aqua
+using SurrogatesBase
+using Aqua
+using JET
+using LinearAlgebra
+using Test
+import Statistics
+
 @testset "Aqua" begin
     Aqua.find_persistent_tasks_deps(SurrogatesBase)
     Aqua.test_ambiguities(SurrogatesBase, recursive = false)
@@ -8,4 +14,97 @@ using SurrogatesBase, Aqua
     Aqua.test_stale_deps(SurrogatesBase)
     Aqua.test_unbound_args(SurrogatesBase)
     Aqua.test_undefined_exports(SurrogatesBase)
+end
+
+# On Julia 1.12+, LinearAlgebra.norm_recursive_check has a type inference issue
+# that causes JET false positives. We ignore LinearAlgebra and Base modules
+# to filter these stdlib issues while still checking our own code.
+const JET_CONFIG = (
+    ignored_modules = (
+        JET.AnyFrameModule(LinearAlgebra),
+        JET.AnyFrameModule(Base),
+    ),
+)
+
+@testset "JET static analysis" begin
+    @testset "Package analysis" begin
+        result = JET.report_package(SurrogatesBase; JET_CONFIG...)
+        @test length(JET.get_reports(result)) == 0
+    end
+
+    @testset "DummySurrogate type stability" begin
+        # Test implementation from runtests.jl
+        struct JETDummySurrogate{X, Y} <: AbstractDeterministicSurrogate
+            xs::Vector{X}
+            ys::Vector{Y}
+        end
+        (s::JETDummySurrogate)(x) = s.ys[argmin([norm(x - ξ) for ξ in s.xs])]
+        function SurrogatesBase.update!(s::JETDummySurrogate, new_xs, new_ys)
+            append!(s.xs, new_xs)
+            append!(s.ys, new_ys)
+        end
+
+        d = JETDummySurrogate(Vector{Vector{Float64}}(), Vector{Int}())
+        SurrogatesBase.update!(d, [[10.3, 0.1], [1.9, 2.1]], [5, 6])
+
+        # Test call method
+        result = JET.report_call(d, Tuple{Vector{Float64}}; JET_CONFIG...)
+        @test length(JET.get_reports(result)) == 0
+
+        # Test update! method
+        result = JET.report_call(
+            SurrogatesBase.update!,
+            Tuple{
+                JETDummySurrogate{Vector{Float64}, Int}, Vector{Vector{Float64}},
+                Vector{Int},
+            }; JET_CONFIG...
+        )
+        @test length(JET.get_reports(result)) == 0
+    end
+
+    @testset "Stochastic surrogate type stability" begin
+        mutable struct JETDummyStochasticSurrogate{X, Y} <: AbstractStochasticSurrogate
+            xs::Vector{X}
+            ys::Vector{Y}
+            ys_mean::Y
+        end
+        function SurrogatesBase.update!(s::JETDummyStochasticSurrogate, new_xs, new_ys)
+            append!(s.xs, new_xs)
+            append!(s.ys, new_ys)
+            s.ys_mean = (s.ys_mean * (length(s.xs) - length(new_xs)) + sum(new_ys)) /
+                length(s.xs)
+        end
+        SurrogatesBase.parameters(s::JETDummyStochasticSurrogate) = s.ys_mean
+
+        struct JETFiniteDummyStochasticSurrogate{X}
+            means::Vector{X}
+        end
+        Statistics.mean(s::JETFiniteDummyStochasticSurrogate) = s.means
+        function JETFiniteDummyStochasticSurrogate(s, xs)
+            return JETFiniteDummyStochasticSurrogate(s.ys_mean .* ones(length(xs)))
+        end
+        function SurrogatesBase.finite_posterior(s::JETDummyStochasticSurrogate, xs)
+            JETFiniteDummyStochasticSurrogate(s, xs)
+        end
+
+        # Test update! method
+        result = JET.report_call(
+            SurrogatesBase.update!,
+            Tuple{
+                JETDummyStochasticSurrogate{Vector{Float64}, Float64},
+                Vector{Vector{Float64}}, Vector{Float64},
+            }; JET_CONFIG...
+        )
+        @test length(JET.get_reports(result)) == 0
+
+        # Test finite_posterior
+        result = JET.report_call(
+            SurrogatesBase.finite_posterior,
+            Tuple{
+                JETDummyStochasticSurrogate{Vector{Float64}, Float64},
+                Vector{Vector{Float64}},
+            }; JET_CONFIG...
+        )
+        @test length(JET.get_reports(result)) == 0
+    end
 end
